+++ src/graphql/user.ts
@@ -43,6 +43,26 @@ export const USER_MEMBER_OF_FRAGMENT = gql`
   }
 `;
 
+export const QUERY_USER_LIST = gql`
+  query users {
+    users {
+      ...UserDetails
+    }
+  }
+  ${USER_DETAILS_FRAGMENT}
+`;
+
+export const QUERY_USER = gql`
+  query user($id: String!) {
+    user(ID: $id) {
+      ...UserDetails
+      ...UserMembers
+    }
+  }
+  ${USER_DETAILS_FRAGMENT}
+  ${USER_MEMBER_OF_FRAGMENT}
+`;
+
 export const QUERY_ECOVERSE_USER_IDS = gql`
   query ecoverseUserIds {
     users {
